task 1 and 2

+++ src/iterator/SortMerge.java
@@ -252,7 +252,7 @@ public class SortMerge extends Iterator implements GlobalConst
 	   Exception
     {
       
-      int    comp_res;
+      double    comp_res;
       Tuple _tuple1,_tuple2;
       if (done) return null;
